alter useAuth
