+++ src/app/components/useAuth.tsx
@@ -1,10 +1,8 @@
 import { useEffect } from "react";
-import { getAuthToken, getMeAuth } from "../services/User/GetAuthToken";
-import { useUserStore } from "../state/userState";
+import { getAuthToken } from "../services/User/GetAuthToken";
 import { logoutUser } from "../services/User/logoutUser";
 
 export function useAuth() {
-  const alterUserCurrent = useUserStore((state) => state.alter);
 
   useEffect(() => {
     const token = getAuthToken();
@@ -12,16 +10,5 @@ export function useAuth() {
       logoutUser();
       return;
     }
-
-    async function fetchUserData() {
-      const userCurrent = await getMeAuth();
-      if (userCurrent) {
-        alterUserCurrent(userCurrent);
-      } else {
-        logoutUser();
-      }
-    }
-
-    fetchUserData();
-  }, [alterUserCurrent]);
+  }, []);
 }
